CRC 64bit fix

+++ include/math/CRC.h
@@ -19,10 +19,12 @@
 #ifndef HPL_CRC_H
 #define HPL_CRC_H
 
+#include <stdint.h>
+
 namespace hpl {
 	//----------------------------------------
 
-	typedef unsigned long tCRCKey;
+	typedef uint32_t tCRCKey;
 
 	//----------------------------------------
 
@@ -30,9 +32,9 @@ namespace hpl {
 	{
 	public:
 		cCRCTable () : mKey (0) {}
-		
+
 		void Init (tCRCKey key);
-		
+
 		tCRCKey operator [] (unsigned i){return mTable [i];}
 
 	private:
@@ -49,7 +51,7 @@ namespace hpl {
 		{
 			mTable.Init (key);
 		}
-		
+
 		void PutByte (unsigned aByte);
 
 		tCRCKey Done ()
@@ -61,7 +63,7 @@ namespace hpl {
 
 	private:
 		static cCRCTable mTable;
-		tCRCKey mKey;	
+		tCRCKey mKey;
 		tCRCKey mRegister;
 	};
